backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
httpx==0.26.0
aiofiles==23.2.1
boto3==1.34.34
python-multipart==0.0.9
argon2-cffi==23.1.0
pyotp==2.9.0
cryptography==42.0.7
webauthn==1.11.1
redis==5.0.8
Pillow>=12.1.0
pillow-heif>=1.2.0
pytest==8.2.2
pytest-asyncio==0.23.7
asgi-lifespan==2.1.0
